buffer_autohide: keep alive visited buffer for a given time
When a buffer has been selected for a reasonable time, keep it visible for an
extra time (set to 1 minute) that is computed when the buffer has been unselected.

Move the logic to check if a buffer should kept visible inside buffer_is_hidable,
while at callbacks we create timeouts to verify if enough time has passed and
in case to hide the buffers.

Fixes #4, improves #3